Quiet some warnings. (vc8 certainly bitches alot more than vc6 did)

+++ stepmania/src/NoteData.cpp
@@ -279,7 +279,7 @@ void NoteData::SetTapAttackNote( int track, int row, Attack attack )
 				TapNote::attack, 
 				TapNote::original, 
 				true,
-				i, 
+				(uint8_t)i, 
 				false,
 				0 );
 			SetTapNote( track, row, tn );
