+++ stepmania/src/PercentageDisplay.cpp
@@ -93,7 +93,7 @@ void PercentageDisplay::Refresh()
 		{		
 			// TRICKY: printf will round, but we want to truncate.  Otherwise, we may display a percent
 			// score that's too high and doesn't match up with the calculated grade.
-			float fTruncInterval = powf(0.1f,PERCENT_TOTAL_SIZE);
+			float fTruncInterval = powf(0.1f, (float) PERCENT_TOTAL_SIZE);
 			fPercentDancePoints = ftruncf( fPercentDancePoints, fTruncInterval );
 			
 			sNumToDisplay = ssprintf( "%*.*f%%", PERCENT_TOTAL_SIZE, PERCENT_DECIMAL_PLACES, fPercentDancePoints*100 );
